Quite a few changes which is the complete game.
Realised that self.num_letters should be assigned to the length of the set of the self.word attribute. This is because sets do not allow duplicates which is what my previous code allowed.

check_letter function completed which checks if the letter is in the word. This was done by enumerating the self.word attribute. Enumerating it allows the code to iterate over the characters in self.word and get the corresponding indices for them. Players num_lives is decreased everytime they guess the incorrect letter.

ask_letter function completed which asks the user for a letter and checks if the letter has already been tried and if the character is a single character. As an extra detail i've also made the code check that the letter isalpha() and will prompt the user to enter a letter if they do not.

play_game function completed which checks whether the player has correctly guessed the letters and will either give the player a congratulations or a you lost print depending on which case is true.

Comments for code all added to explain my thought process too.