some small changes, just made the code more accurate to what was requested in the template.
Have added a section of code that decrements the number of unique letters in the mystery word by 1 every time 1 is succesfully guessed.

Code does not account for cases where the unique letter appears twice and will decrease the number of unique letters by 2. Not sure what to do here.